добавили админ страницы

+++ src/main/java/ru/krylov/web/dao/RoleDAO.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface RoleDAO {
 	Role getRoleByName(String name);
 
-	Role getRoleById(Long id);
+	Role getRoleById(int id);
 
 	List<Role> allRoles();
