+++ src/main/java/ru/krylov/web/service/RoleService.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface RoleService {
 	Role getRoleByName(String name);
 
-	Role getRoleById(Long id);
+	Role getRoleById(int id);
 
 	List<Role> allRoles();
